Bugfixes, new pic.
