+++ src/worldgen/biome_desert.h
@@ -1,0 +1,8 @@
+#ifndef _biome_desert_h_
+#define _biome_desert_h_
+
+#include "world.h"
+
+void generateDesert(int dx, int dz, int x, int z, int h, int flag, world_func func, void *arg);
+
+#endif // _biome_desert_h_
