+++ src/worldgen/biome_rainforest.h
@@ -1,0 +1,8 @@
+#ifndef _biome_rainforest_h_
+#define _biome_rainforest_h_
+
+#include "world.h"
+
+void generateRainforest(int dx, int dz, int x, int z, int h, int flag, world_func func, void *arg);
+
+#endif // _biome_rainforest_h_
