+++ src/worldgen/biome_taiga.h
@@ -1,0 +1,8 @@
+#ifndef _biome_taiga_h_
+#define _biome_taiga_h_
+
+#include "world.h"
+
+void generateTaiga(int dx, int dz, int x, int z, int h, int flag, world_func func, void *arg);
+
+#endif // _biome_taiga_h_
