+++ src/worldgen/biome_temperate.h
@@ -1,0 +1,8 @@
+#ifndef _biome_temperate_h_
+#define _biome_temperate_h_
+
+#include "world.h"
+
+void generateTemperate(int dx, int dz, int x, int z, int h, int flag, world_func func, void *arg);
+
+#endif // _biome_temperate_h_
